Added fancybox for photo navigation.
URL: http://fancyapps.com/fancybox/

Before this change, image in the gallery were being opened in their own page, without any way to navigate back and forth.

By using fancybox, pictures are automatically resized to current viewport and the gallery can be browsed with the on-screen buttons or by pressing left/right on the keyboard.